# Speed up encodeBinaryNode in x100 times if encoding many items in content