Speed up encodeBinaryNode in **x100 times** if encoding many items in content

The problem was that when we use `encodeBinaryNode(item, opts, buffer)` in for loop we actually don't use the result, but we DO convert it to `Buffer.from`

Here:
https://github.com/WhiskeySockets/Baileys/blob/f25f83656a7cadbcdaf66b7fea72b9d4b3a72aa0/src/WABinary/encode.ts#L227

I've moved `Buffer.from` to `sendNode` function, so it converts the array to Buffer only when required (at the end), but not in the middle of the inner loop


```
encodeBinaryNode NEW - 1 - 74ms
encodeBinaryNode NEW - 2 - 68ms
encodeBinaryNode NEW - 3 - 88ms
encodeBinaryNode NEW - 4 - 45ms
encodeBinaryNode NEW - 5 - 34ms
encodeBinaryNode NEW - 6 - 33ms
encodeBinaryNode NEW - 7 - 38ms
encodeBinaryNode NEW - 8 - 30ms
encodeBinaryNode NEW - 9 - 31ms
encodeBinaryNode ORIGINAL - 1 - 4846ms
encodeBinaryNode ORIGINAL - 2 - 4562ms
# I didn't wait the original 10 times, too long
```

With the fix it took **3 seconds**  to send status to **3K contacts** (including IO communication staff with store and WA servers)