@devlikepro 2024-06-21
If anyone find that the fix doesn't improve their speed - could you collect your `{ tag, attrs, content }` for that case and share it with me so I can improve it? :pray: