@Salientekill 2024-06-21
Thanks for the work, I'm looking forward to seeing improvements in terms of speed