@devlikepro 2024-06-26
@PurpShell could you look at this when you have time? :pray: 
It's just basically moving Buffer.from to different place, nothing serious (but the impact IS)