@devlikepro 2024-07-03
> the fact that encodeBinaryNode is nested and iterates itself and its children

and that it's not used later, correct. 

> potential breaking change for anyone using those functions themselves

Oh, I see. We could add two different method - one for inner `encodeBinaryNodeInner` that would always return array and one for `encodeBinaryNode` that would always return Buffer, so no external users will be affected and no Buffer.from will be in `socket.ts`

Let me know if you prefer that solution!